requirements.txt
argcomplete
azure-devops
PyGithub
PyYAML
